feature[ats_ATS146075]: Implement & Integrate ORCS Indexer JAX-RS API
Implement ORCS Indexer JAX-RS API
Integrate ORCS Indexer JAX-RS API
Remove SearchEngineTaggerServlet

+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/transaction/StoreSkynetTransactionOperation.java
@@ -122,14 +122,14 @@ public final class StoreSkynetTransactionOperation extends AbstractDbTxOperation
    }
 
    private void tagGammas() throws OseeCoreException {
-      Set<Integer> gammasToTag = new LinkedHashSet<Integer>();
+      Set<Long> gammasToTag = new LinkedHashSet<Long>();
       for (BaseTransactionData transactionData : txDatas) {
          if (!transactionData.getModificationType().isExistingVersionUsed() && transactionData instanceof AttributeTransactionData) {
             AttributeTransactionData attrData = (AttributeTransactionData) transactionData;
             if (!attrData.getAttribute().isUseBackingData()) {
                Attribute<?> attr = ((AttributeTransactionData) transactionData).getAttribute();
                if (attr.getAttributeType().isTaggable()) {
-                  gammasToTag.add(transactionData.getGammaId());
+                  gammasToTag.add((long) transactionData.getGammaId());
                }
             }
          }
